+++ src/configuration/models/spdx/ISpdx.ts
@@ -1,0 +1,8 @@
+/**
+ * Model of spdx.json file.
+ */
+import { ISpdxLicense } from "./ISpdxLicense";
+
+export interface ISpdx {
+    [id: string]: ISpdxLicense;
+}