+++ src/configuration/models/unite/uniteConfiguration.ts
@@ -5,6 +5,7 @@ import { IncludeMode } from "./includeMode";
 import { UniteDirectories } from "./uniteDirectories";
 import { UniteLinter } from "./uniteLinter";
 import { UniteModuleLoader } from "./uniteModuleLoader";
+import { UnitePackageManager } from "./unitePackageManager";
 import { UniteSourceLanguage } from "./uniteSourceLanguage";
 import { UniteUnitTestFramework } from "./uniteUnitTestFramework";
 import { UniteUnitTestRunner } from "./uniteUnitTestRunner";
@@ -12,9 +13,11 @@ import { UniteUnitTestRunner } from "./uniteUnitTestRunner";
 export class UniteConfiguration {
     public packageName: string;
     public title: string;
+    public license: string;
     public sourceLanguage: UniteSourceLanguage;
     public moduleLoader: UniteModuleLoader;
     public linter: UniteLinter;
+    public packageManager: UnitePackageManager;
     public unitTestRunner: UniteUnitTestRunner;
     public unitTestFramework: UniteUnitTestFramework;