+++ src/configuration/models/unite/uniteDependencies.ts
@@ -1,6 +1,0 @@
-/**
- * Model of Unite Dependencies (unite-dependencies.json) file.
- */
-export class UniteDependencies {
-    public versions: { [id: string]: string };
-}