+++ src/configuration/models/unite/unitePackageManager.ts
@@ -1,0 +1,6 @@
+/**
+ * Type for Unite Configuration packageManager.
+ */
+export type UnitePackageManager =
+    "Npm" |
+    "Yarn";