+++ src/engine/enginePipelineStepBase.ts
@@ -2,7 +2,6 @@
  * Base implementation of engine pipeline step.
  */
 import { UniteConfiguration } from "../configuration/models/unite/uniteConfiguration";
-import { ErrorHandler } from "../core/errorHandler";
 import { IDisplay } from "../interfaces/IDisplay";
 import { IEnginePipelineStep } from "../interfaces/IEnginePipelineStep";
 import { IFileSystem } from "../interfaces/IFileSystem";
@@ -13,18 +12,13 @@ export abstract class EnginePipelineStepBase implements IEnginePipelineStep {
     public abstract process(logger: ILogger, display: IDisplay, fileSystem: IFileSystem, uniteConfiguration: UniteConfiguration, engineVariables: EngineVariables): Promise<number>;
 
     public log(logger: ILogger, display: IDisplay, message: string, args?: { [id: string]: any}): void {
-        display.log(message, this.arrayToReadable(args));
+        display.log(message, args);
         logger.log(message, args);
     }
 
     public error(logger: ILogger, display: IDisplay, message: string, err: any, args?: { [id: string]: any}): void {
-        if (err) {
-            display.error(message + ": " + ErrorHandler.format(err));
-            logger.exception(message, err, args);
-        } else {
-            display.error(message + ": " + this.arrayToReadable(args));
-            logger.error(message, args);
-        }
+        display.error(message, err, args);
+        logger.error(message, args);
     }
 
     public async copyFile(logger: ILogger, display: IDisplay, fileSystem: IFileSystem,
@@ -33,14 +27,5 @@ export abstract class EnginePipelineStepBase implements IEnginePipelineStep {
 
         const lines = await fileSystem.fileReadLines(sourceFolder, sourceFilename);
         await fileSystem.fileWriteLines(destFolder, destFilename, lines);
-    }
-
-    private arrayToReadable(args?: { [id: string]: any}): string {
-        if (!args) {
-            return "";
-        } else {
-            const objKeys = Object.keys(args);
-            return (objKeys.length === 0 ? "" : (objKeys.length === 1 ? args[objKeys[0]] : JSON.stringify(args)));
-        }
     }
 }