+++ src/interfaces/IEngine.ts
@@ -4,11 +4,13 @@
 export interface IEngine {
     init(packageName: string | undefined | null,
          title: string | undefined | null,
+         license: string | undefined | null,
          sourceLanguage: string | undefined | null,
          moduleLoader: string | undefined | null,
          unitTestRunner: string | undefined | null,
          unitTestFramework: string | undefined | null,
          linter: string | undefined | null,
+         packageManager: string | undefined | null,
          outputDirectory: string | undefined | null): Promise<number>;
 
     clientPackage(operation: string | undefined | null,
@@ -16,5 +18,6 @@ export interface IEngine {
                   version: string | undefined | null,
                   preload: boolean,
                   includeMode: string | undefined | null,
+                  packageManager: string | undefined | null,
                   outputDirectory: string | undefined | null): Promise<number>;
 }