+++ src/interfaces/ILogger.ts
@@ -4,6 +4,5 @@
 export interface ILogger {
     log(message: string, args?: { [id: string]: any }): void;
     info(message: string, args?: { [id: string]: any }): void;
-    error(message: string, args?: { [id: string]: any }): void;
-    exception(message: string, exception: any, args?: { [id: string]: any }): void;
+    error(message: string, exception: any, args?: { [id: string]: any }): void;
 }