+++ src/pipelineSteps/content/gitIgnore.ts
@@ -1,12 +1,12 @@
 /**
  * Pipeline step to generate .gitignore.
  */
-import { UniteConfiguration } from "../configuration/models/unite/uniteConfiguration";
-import { EnginePipelineStepBase } from "../engine/enginePipelineStepBase";
-import { EngineVariables } from "../engine/engineVariables";
-import { IDisplay } from "../interfaces/IDisplay";
-import { IFileSystem } from "../interfaces/IFileSystem";
-import { ILogger } from "../interfaces/ILogger";
+import { UniteConfiguration } from "../../configuration/models/unite/uniteConfiguration";
+import { EnginePipelineStepBase } from "../../engine/enginePipelineStepBase";
+import { EngineVariables } from "../../engine/engineVariables";
+import { IDisplay } from "../../interfaces/IDisplay";
+import { IFileSystem } from "../../interfaces/IFileSystem";
+import { ILogger } from "../../interfaces/ILogger";
 
 export class GitIgnore extends EnginePipelineStepBase {
     public async process(logger: ILogger, display: IDisplay, fileSystem: IFileSystem, uniteConfiguration: UniteConfiguration, engineVariables: EngineVariables): Promise<number> {