+++ src/pipelineSteps/lint/tsLint.ts
@@ -21,7 +21,18 @@ export class TsLint extends EnginePipelineStepBase {
                 engineVariables.requiredDevDependencies.push("gulp-tslint");
                 engineVariables.requiredDevDependencies.push("tslint");
 
-                const config = this.generateConfig(fileSystem, uniteConfiguration, engineVariables);
+                let existing;
+                try {
+                    const exists = await fileSystem.fileExists(engineVariables.rootFolder, "tslint.json");
+                    if (exists) {
+                        existing = await fileSystem.fileReadJson<TsLintConfiguration>(engineVariables.rootFolder, "tslint.json");
+                    }
+                } catch (err) {
+                    super.error(logger, display, "Reading existing tslint.json failed", err);
+                    return 0;
+                }
+
+                const config = this.generateConfig(fileSystem, uniteConfiguration, engineVariables, existing);
                 await fileSystem.fileWriteJson(engineVariables.rootFolder, "tslint.json", config);
 
                 return 0;
@@ -44,12 +55,18 @@ export class TsLint extends EnginePipelineStepBase {
         return 0;
     }
 
-    private generateConfig(fileSystem: IFileSystem, uniteConfiguration: UniteConfiguration, engineVariables: EngineVariables): TsLintConfiguration {
+    private generateConfig(fileSystem: IFileSystem, uniteConfiguration: UniteConfiguration, engineVariables: EngineVariables, existing: TsLintConfiguration | undefined): TsLintConfiguration {
         const config = new TsLintConfiguration();
 
         config.extends = "tslint:recommended";
         config.rulesDirectory = [];
         config.rules = {};
+
+        if (existing) {
+            config.extends = existing.extends || config.extends;
+            config.rulesDirectory = existing.rulesDirectory || config.rulesDirectory;
+            config.rules = existing.rules || config.rules;
+        }
 
         return config;
     }