+++ src/pipelineSteps/packageJson.ts
@@ -3,7 +3,6 @@
  */
 import { PackageConfiguration } from "../configuration/models/packages/packageConfiguration";
 import { UniteConfiguration } from "../configuration/models/unite/uniteConfiguration";
-import { UniteDependencies } from "../configuration/models/unite/uniteDependencies";
 import { EnginePipelineStepBase } from "../engine/enginePipelineStepBase";
 import { EngineVariables } from "../engine/engineVariables";
 import { IDisplay } from "../interfaces/IDisplay";
@@ -12,14 +11,14 @@ import { ILogger } from "../interfaces/ILogger";
 
 export class PackageJson extends EnginePipelineStepBase {
     public async process(logger: ILogger, display: IDisplay, fileSystem: IFileSystem, uniteConfiguration: UniteConfiguration, engineVariables: EngineVariables): Promise<number> {
-        let uniteDependencies: UniteDependencies;
+        let corePackageJson: PackageConfiguration;
 
         try {
-            super.log(logger, display, "Loading dependencies", { assetsDirectory: engineVariables.assetsDirectory, dependenciesFile: engineVariables.dependenciesFile });
+            super.log(logger, display, "Loading dependencies", { core: engineVariables.coreFolder, dependenciesFile: "package.json" });
 
-            uniteDependencies = await fileSystem.fileReadJson<UniteDependencies>(engineVariables.assetsDirectory, engineVariables.dependenciesFile);
+            corePackageJson = await fileSystem.fileReadJson<PackageConfiguration>(engineVariables.coreFolder, "package.json");
         } catch (err) {
-            super.error(logger, display, "Loading dependencies failed", err, { assetsDirectory: engineVariables.assetsDirectory, dependenciesFile: engineVariables.dependenciesFile });
+            super.error(logger, display, "Loading dependencies failed", err, { core: engineVariables.coreFolder, dependenciesFile: "package.json" });
             return 1;
         }
 
@@ -29,8 +28,9 @@ export class PackageJson extends EnginePipelineStepBase {
             const packageJson = new PackageConfiguration();
             packageJson.name = uniteConfiguration.packageName;
             packageJson.version = "0.0.1";
-            packageJson.dependencies = this.lookupDependencies(logger, display, engineVariables.requiredDependencies, uniteDependencies);
-            packageJson.devDependencies = this.lookupDependencies(logger, display, engineVariables.requiredDevDependencies, uniteDependencies);
+            packageJson.license = uniteConfiguration.license;
+            packageJson.dependencies = this.lookupDependencies(logger, display, engineVariables.requiredDependencies, corePackageJson);
+            packageJson.devDependencies = this.lookupDependencies(logger, display, engineVariables.requiredDevDependencies, corePackageJson);
 
             if (uniteConfiguration.clientPackages) {
                 for (const pkg in uniteConfiguration.clientPackages) {
@@ -50,16 +50,16 @@ export class PackageJson extends EnginePipelineStepBase {
         }
     }
 
-    private lookupDependencies(logger: ILogger, display: IDisplay, requiredDependencies: string[], uniteDependencies: UniteDependencies): { [id: string]: string} {
+    private lookupDependencies(logger: ILogger, display: IDisplay, requiredDependencies: string[], corePackageJson: PackageConfiguration): { [id: string]: string} {
         const dependencies: { [id: string]: string } = {};
 
         if (requiredDependencies) {
-            if (uniteDependencies && uniteDependencies.versions) {
+            if (corePackageJson && corePackageJson.peerDependencies) {
                 requiredDependencies.sort();
 
                 requiredDependencies.forEach(requiredDependency => {
-                    if (uniteDependencies.versions[requiredDependency]) {
-                        dependencies[requiredDependency] = uniteDependencies.versions[requiredDependency];
+                    if (corePackageJson.peerDependencies[requiredDependency]) {
+                        dependencies[requiredDependency] = corePackageJson.peerDependencies[requiredDependency];
                     } else {
                         throw new Error("Missing Dependency '" + requiredDependency + "'");
                     }